# 06_api_reference/openapi_spec.yaml
...
openapi: 3.1.0
info:
  title: Teivaka TFOS API
  description: |
    Teivaka Agricultural Transformation Operating System (Agri-TOS) API.

    **Platform:** Teivaka PTE LTD (Company No. 2025RC001894), Fiji
    **Founder:** Uraia Koroi Kama (Cody)

    ## Four Pillars
    - **Knowledge Base (KB):** Validated agronomic protocols — RAG-grounded, no hallucination
    - **Farm OS (TFOS):** Production cycles, harvests, labor, inputs, tasks, alerts
    - **AI Intelligence (TIS):** Teivaka Intelligence System — Voice + Text AI for farmers
    - **Community:** Shared protocols and benchmarks across Teivaka network

    ## Authentication
    - Bearer JWT (15-minute access token)
    - httpOnly refresh cookie (7-day refresh token)
    - Multi-tenancy: All operational endpoints enforce `tenant_id` via PostgreSQL RLS

    ## Primary Financial Metric
    **CoKG** (Cost per Kilogram) = (LaborCost + InputCost + OtherCost) / TotalHarvestQty_kg
    CoKG is always the **first field** in financial response objects.

    ## Fiji-Specific
    - Currency: FJD (Fijian Dollar), all monetary fields suffixed `_fjd`
    - Timezone: Pacific/Fiji (UTC+12)
    - WhatsApp-first communication via Twilio
    - Voice pipeline: Audio → Whisper → TIS → TFOS → Confirm (target: <5 seconds)

    ## TIS Rate Limits (Redis-enforced)
    | Tier | TIS Calls/Day |
    |------|---------------|
    | FREE | 5 |
    | BASIC | 20 |
    | PREMIUM | Unlimited |
    | CUSTOM | Unlimited |

    ## Real Farm Data (Reference)
...

...
servers:
...
security:
  - BearerAuth: []

# ============================================================
# COMPONENTS
# ============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (15-minute expiry). Obtain from POST /auth/login.
        Format: Authorization: Bearer <token>
...
paths:
...
  /farms/{farm_id}/alerts:
    get:
      tags: [Alerts]
      summary: List farm alerts
      security:
        - BearerAuth: []
      parameters:
        - name: farm_id
          in: path
          required: true
          schema:
            type: string
          example: F001
        - name: severity
          in: query
          schema:
            type: string
            enum: [Critical, High, Medium, Low]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved]
            default: open
        - name: pu_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              example:
                success: true
                data:
                  - alert_id: ALT-20260403-001
                    rule_id: RULE-038
                    title: Chemical Compliance Block
                    description: "Dimethoate (CHEM-001) applied 4 days ago. Safe to harvest after 2026-04-10."
                    severity: Critical
                    rag_color: RED
                    status: open
                    pu_id: F001-PU002
                    cycle_id: CY-F001-26-002
                    created_at: "2026-04-03T08:00:00+12:00"
  /alerts/{alert_id}/acknowledge:
    put:
      tags: [Alerts]
      summary: Acknowledge an alert
      security:
        - BearerAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
          example: ALT-20260403-001
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              example:
                success: true
                data:
                  alert_id: ALT-20260403-001
                  status: acknowledged
                  acknowledged_at: "2026-04-07T09:30:00+12:00"
                  acknowledged_by: USR-001
  /alerts/{alert_id}/resolve:
    put:
      tags: [Alerts]
      summary: Manually resolve an alert (FOUNDER/MANAGER)
      security:
        - BearerAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
          example: ALT-20260403-001
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                resolution_notes:
                  type: string
      responses:
        '200':
          description: Alert resolved
          content:
            application/json:
              example:
                success: true
                data:
                  alert_id: ALT-20260403-001
                  status: resolved
                  resolved_at: "2026-04-07T10:00:00+12:00"

  # ==================== TASKS ====================
...